+++ HISTORY.md
@@ -13,6 +13,7 @@ http://visjs.org
 
 - Fixed error with zero nodes with hierarchical layout.
 - Added focusOnNode function.
+- Added Hover option.
 
 
 ## 2014-05-28, version 1.0.2